feat: start hooking up finished events so we know when modes or heads are dropped
Still some work to be done on this front, such as re-applying configuration dynamically

feat: changed refresh rate to Hz for better precision

+++ src/config/format.hpp
@@ -33,7 +33,7 @@ struct DisplayGroupOutputConfig {
     std::string        serial;
     int                width;
     int                height;
-    double             refresh;
+    int                refresh;
     std::array<int, 2> position;
     double             scale;
     int                rotation;
